Import latest dsub changes

+++ dsub/providers/local.py
@@ -55,6 +55,7 @@ from datetime import datetime
 from datetime import timedelta
 import os
 import signal
+import string
 import subprocess
 import tempfile
 import textwrap
@@ -110,12 +111,49 @@ _SUPPORTED_INPUT_PROVIDERS = _SUPPORTED_FILE_PROVIDERS
 _SUPPORTED_OUTPUT_PROVIDERS = _SUPPORTED_FILE_PROVIDERS
 
 
+def _format_task_name(job_id, task_id):
+  """Create a task name from a job-id and a task-id.
+
+  Task names are used internally by dsub as well as by the docker task runner.
+  The name is formatted as either "<job-id>.<task-id>" or jobs with multple
+  tasks, or just "<job-id>" for jobs with a single task. Task names follow
+  formatting conventions allowing them to be safely used as a docker name.
+
+  Args:
+    job_id: (str) the job ID.
+    task_id: (str) the task ID.
+
+  Returns:
+    a task name string.
+  """
+  if task_id is None:
+    docker_name = job_id
+  else:
+    docker_name = '%s.%s' % (job_id, task_id)
+
+  # Docker container names must match: [a-zA-Z0-9][a-zA-Z0-9_.-]
+  # So 1) prefix it with "dsub-" and 2) change all invalid characters to "-".
+  return 'dsub-{}'.format(_convert_suffix_to_docker_chars(docker_name))
+
+
+def _convert_suffix_to_docker_chars(suffix):
+  """Rewrite string so that all characters are valid in a docker name suffix."""
+  # Docker container names must match: [a-zA-Z0-9][a-zA-Z0-9_.-]
+  accepted_characters = string.ascii_letters + string.digits + '_.-'
+
+  def label_char_transform(char):
+    if char in accepted_characters:
+      return char
+    return '-'
+
+  return ''.join(label_char_transform(c) for c in suffix)
+
+
 class LocalJobProvider(base.JobProvider):
   """Docker jobs running locally (i.e. on the caller's computer)."""
 
   def __init__(self):
     self._operations = []
-    self.provider_root_cache = None
 
   def prepare_job_metadata(self, script, job_name, user_id):
     job_name_value = job_name or os.path.basename(script)
@@ -213,8 +251,8 @@ class LocalJobProvider(base.JobProvider):
       }}
 
       localize_data() {{
-        true # ensure body is not empty, to avoid error.
         {localize_command}
+        recursive_localize_data
       }}
 
       recursive_delocalize_data() {{
@@ -223,33 +261,39 @@ class LocalJobProvider(base.JobProvider):
       }}
 
       delocalize_data() {{
-        true # ensure body is not empty, to avoid error.
         {delocalize_command}
+        recursive_delocalize_data
       }}
 
       delocalize_logs() {{
-        true # ensure body is not empty, to avoid error.
         {delocalize_logs_command}
       }}
       """)
     script_body = textwrap.dedent("""\
       # Delete local files
       cleanup() {
+        local rm_data_dir="${1:-true}"
+
         echo "Copying the logs before cleanup"
         delocalize_logs
         # Not putting it in status_message because we don't want that to be the
         # last line in case of error.
-        echo "cleaning up ${DATA_DIR}"
-        # Clean up files written from inside Docker
-        2>&1 docker run \\
-         --name "${NAME}-cleanup" \\
-         --workdir "${DATA_MOUNT_POINT}/${WORKING_DIR}" \\
-         "${VOLUMES[@]}" \\
-         --env-file "${ENV_FILE}" \\
-         "${IMAGE}" \\
-         rm -rf "${DATA_MOUNT_POINT}/*" | tee -a log.txt
+
         # Clean up files staged from outside Docker
-        rm -rf "${DATA_DIR}" || echo "sorry, unable to delete ${DATA_DIR}."
+        if [[ "${rm_data_dir}" == "true" ]]; then
+          echo "cleaning up ${DATA_DIR}"
+
+          # Clean up files written from inside Docker
+          2>&1 docker run \\
+            --name "${NAME}-cleanup" \\
+            --workdir "${DATA_MOUNT_POINT}/${WORKING_DIR}" \\
+            "${VOLUMES[@]}" \\
+            --env-file "${ENV_FILE}" \\
+            "${IMAGE}" \\
+            rm -rf "${DATA_MOUNT_POINT}/*" | tee -a log.txt
+
+          rm -rf "${DATA_DIR}" || echo "sorry, unable to delete ${DATA_DIR}."
+        fi
       }
 
       log_info() {
@@ -270,21 +314,18 @@ class LocalJobProvider(base.JobProvider):
         local parent_lineno="$1"
         local code="$2"
         local message="${3:-Error}"
+
+        # Disable further traps
+        trap EXIT
+        trap ERR
+
         if [[ $code != "0" ]]; then
           echo "FAILURE" > status.txt
           log_error "${message} on or near line ${parent_lineno}; exiting with status ${code}"
         fi
-        cleanup
-        # Disable further traps
-        trap EXIT
+        cleanup "false"
         exit "${code}"
       }
-      # This will trigger whenever a command returns an error code
-      # (exactly like set -e)
-      trap 'error ${LINENO} $? Error' ERR
-      # This will trigger on all other exits. We disable it before normal
-      # exit so we know if it fires it means there's a problem.
-      trap 'error ${LINENO} $? "Exit (undefined variable or kill?)"' EXIT
 
       fetch_image() {
         local image="$1"
@@ -337,20 +378,30 @@ class LocalJobProvider(base.JobProvider):
           chown -R "${usergroup}" "${docker_directory}" >> stdout.txt 2>> stderr.txt
       }
 
-
       exit_if_canceled() {
         if [[ -f die ]]; then
           log_info "Job is canceled, stopping Docker container ${NAME}."
           docker stop "${NAME}"
           echo "CANCELED" > status.txt
           log_info "Delocalize logs and cleanup"
-          cleanup
+          cleanup "false"
           trap EXIT
           echo "Canceled, exiting." > status_message.txt
           exit 1
         fi
       }
 
+
+      # This will trigger whenever a command returns an error code
+      # (exactly like set -e)
+      trap 'error ${LINENO} $? Error' ERR
+
+      # This will trigger on all other exits. We disable it before normal
+      # exit so we know if it fires it means there's a problem.
+      trap 'error ${LINENO} $? "Exit (undefined variable or kill?)"' EXIT
+
+      # Make sure that ERR traps are inherited by shell functions
+      set -o errtrace
 
       # Beginning main execution
 
@@ -417,7 +468,7 @@ class LocalJobProvider(base.JobProvider):
       # recursively)
       trap EXIT
       log_info "Delocalize logs and cleanup."
-      cleanup
+      cleanup "true"
       if [[ -z "${FAILURE_MESSAGE}" ]]; then
         echo "SUCCESS" > status.txt
         log_info "Done"
@@ -435,7 +486,7 @@ class LocalJobProvider(base.JobProvider):
 
     script = script_header.format(
         volumes=volumes,
-        name=LocalTask.format_docker_name(
+        name=_format_task_name(
             task_metadata.get('job-id'), task_metadata.get('task-id')),
         image=job_resources.image,
         script=DATA_MOUNT_POINT + '/' + SCRIPT_DIR + '/' +
@@ -589,7 +640,6 @@ class LocalJobProvider(base.JobProvider):
     task_list = None if task_list == ['*'] else task_list
     # 'AND' filtering arguments.
     labels = labels if labels else []
-
 
     create_time_local = self._utc_int_to_local_datetime(create_time)
 
@@ -767,9 +817,7 @@ class LocalJobProvider(base.JobProvider):
         pid=pid)
 
   def _provider_root(self):
-    if not self.provider_root_cache:
-      self.provider_root_cache = tempfile.gettempdir() + '/dsub-local'
-    return self.provider_root_cache
+    return tempfile.gettempdir() + '/dsub-local'
 
   def _delocalize_logging_command(self, file_provider, task_metadata):
     """Returns a command to delocalize logs.
@@ -885,20 +933,24 @@ class LocalJobProvider(base.JobProvider):
     for i in ins:
       if i.recursive:
         continue
-      gcs_file_path = i.value
-      local_file_path = task_dir + '/' + DATA_SUBDIR + '/' + i.docker_path
-      commands.append('mkdir -p "%s"' % os.path.dirname(local_file_path))
-      commands.append('gsutil -q cp "%s" "%s"' %
-                      (gcs_file_path,
-                       self._get_input_target_path(local_file_path)))
 
-    commands.append(
-        textwrap.dedent("""\
-    if ! recursive_localize_data; then
-      log_error "Recursive localization failed."
-      exit 1
-    fi
-    """))
+      source_file_path = i.uri
+      local_file_path = task_dir + '/' + DATA_SUBDIR + '/' + i.docker_path
+      dest_file_path = self._get_input_target_path(local_file_path)
+
+      commands.append('mkdir -p "%s"' % os.path.dirname(local_file_path))
+
+      if i.file_provider in [param_util.P_LOCAL, param_util.P_GCS]:
+        # The semantics that we expect here are implemented consistently in
+        # "gsutil cp", and are a bit different than "cp" when it comes to
+        # wildcard handling, so use it for both local and GCS:
+        #
+        # - `cp path/* dest/` will error if "path" has subdirectories.
+        # - `cp "path/*" "dest/"` will fail (it expects wildcard expansion
+        #   to come from shell).
+        commands.append('gsutil -q cp "%s" "%s"' % (source_file_path,
+                                                    dest_file_path))
+
     return '\n'.join(commands)
 
   def _mkdir_outputs(self, task_dir, task_data):
@@ -934,20 +986,19 @@ class LocalJobProvider(base.JobProvider):
     for o in outs:
       if o.recursive:
         continue
+
+      # The destination path is o.uri.path, which is the target directory
+      # (rather than o.uri, which includes the filename or wildcard).
       dest_path = o.uri.path
       local_path = task_dir + '/' + DATA_SUBDIR + '/' + o.docker_path
-      if o.file_provider == param_util.P_GCS:
-        commands.append('gsutil -q cp "%s" "%s"' % (local_path, dest_path))
+
       if o.file_provider == param_util.P_LOCAL:
         commands.append('mkdir -p "%s"' % dest_path)
-        commands.append('cp %s %s' % (local_path, dest_path))
 
-    commands.append(
-        textwrap.dedent("""\
-    if ! recursive_delocalize_data; then
-      log_error "Recursive delocalization failed."
-      exit 1
-    fi"""))
+      # Use gsutil even for local files (explained in _localize_inputs_command).
+      if o.file_provider in [param_util.P_LOCAL, param_util.P_GCS]:
+        commands.append('gsutil -q cp "%s" "%s"' % (local_path, dest_path))
+
     return '\n'.join(commands)
 
   def _stage_script(self, task_dir, script_name, script_text):
@@ -1013,24 +1064,16 @@ class LocalTask(base.Task):
         return self._last_line(tad.get('status-message', None))
     if field == 'status-detail':
       return tad.get('status-message', None)
+    if field == 'start-time':
+      # There's no delay between creation and start since we launch docker
+      # immediately for local runs.
+      return tad.get('create-time', None)
 
     return tad.get(field, default)
 
   def get_docker_name_for_task(self):
-    return self.format_docker_name(
-        self.get_field('job-id'),
-        self.get_field('task-id'))
-
-  @staticmethod
-  def format_docker_name(job_id, task_id):
-    # The name of the docker container is formatted as either:
-    #  <job-id>.<task-id>
-    # for "task" jobs, or just <job-id> for non-task jobs
-    # (those have "None" as the task ID).
-    if task_id is None:
-      return job_id
-    else:
-      return '%s.%s' % (job_id, task_id)
+    return _format_task_name(
+        self.get_field('job-id'), self.get_field('task-id'))
 
   def get_task_status_message(self):
     status = self.get_field('status')
@@ -1046,7 +1089,6 @@ class LocalTask(base.Task):
     if value.endswith('\n'):
       return value.split('\n')[-2]
     return value.split('\n')[-1]
-
 
 if __name__ == '__main__':
   pass
